backend/requirements.txt
# Core dependencies
langchain-community>=0.0.10
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pypdf>=5.0.0

# Vector store
chromadb>=0.5.0

# API and Web
flask>=2.0.0
flask-cors>=4.0.0

# Embeddings and ML
numpy>=1.24.0
sentence-transformers>=2.2.2

# LLM Integration
together>=0.2.5

# Utilities
tqdm>=4.65.0